+++ kontactpickerlib/src/main/java/com/deepakkumardk/kontactpickerlib/KontactPicker.kt
@@ -2,7 +2,6 @@ package com.deepakkumardk.kontactpickerlib
 
 import android.app.Activity
 import android.content.Intent
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import com.deepakkumardk.kontactpickerlib.model.KontactPickerItem
 import com.deepakkumardk.kontactpickerlib.model.MyContacts
@@ -16,8 +15,6 @@ import com.deepakkumardk.kontactpickerlib.util.KontactPickerUI
 class KontactPicker {
 
     fun startPickerForResult(activity: Activity?, item: KontactPickerItem, requestCode: Int) {
-        item.textBgColor =
-            activity?.applicationContext.let { ContextCompat.getColor(it!!, item.textBgColor) }
         KontactPickerUI.setPickerUI(item)
         val intent = Intent(activity, KontactPickerActivity::class.java)
         activity.let {
@@ -26,13 +23,6 @@ class KontactPicker {
     }
 
     fun startPickerForResult(fragment: Fragment?, item: KontactPickerItem, requestCode: Int) {
-        item.textBgColor =
-            fragment?.context.let {
-                ContextCompat.getColor(
-                    it!!,
-                    item.textBgColor
-                )
-            }
         KontactPickerUI.setPickerUI(item)
         val intent = Intent(fragment?.context, KontactPickerActivity::class.java)
         fragment.let {