+++ kontactpickerlib/src/main/java/com/deepakkumardk/kontactpickerlib/model/KontactPickerItem.kt
@@ -5,8 +5,9 @@ package com.deepakkumardk.kontactpickerlib.model
  */
 
 class KontactPickerItem {
-    var textBgColor = 0
     var debugMode = false
+    var textBgColor: Int? = null
+    var themeResId: Int? = null
     var imageMode: ImageMode = ImageMode.None
     var selectionTickView: SelectionTickView = SelectionTickView.SmallView
 }